Shared OpenAI Client Factory
